requirements.txt
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
click==8.0.3
cycler==0.11.0
Cython==0.29.32
debugpy==1.5.1
decorator==5.1.1
dill==0.3.5.1
entrypoints==0.4
executing==0.8.2
fonttools==4.29.1
greenlet==1.1.3
h5py==3.7.0
importlib-metadata==4.12.0
ipykernel==6.8.0
ipython==8.0.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.1
kiwisolver==1.3.2
llvmlite==0.39.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
multiprocess==0.70.13
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numba==0.56.2
numpy==1.22.2
packaging==21.3
pandas==1.4.0
parso==0.8.3
pathos==0.2.9
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
pox==0.3.1
ppft==1.7.6.5
prompt-toolkit==3.0.26
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
Pygments==2.11.2
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
SciencePlots==1.0.9
scikit-learn==1.1.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.41
stack-data==0.1.4
syndp==0.1.0
threadpoolctl==3.1.0
tomli==2.0.0
tornado==6.1
traitlets==5.1.1
tslearn==0.5.2
typing_extensions==4.0.1
wcwidth==0.2.5
zipp==3.8.1
